Status is logged and exception thrown if not status ok
