+++ mimosa/core/solver.py
@@ -4,6 +4,8 @@ from mimosa.common import (
     SolverManagerFactory,
     SolverStatus,
     OptSolver,
+    logger,
+    utils,
 )
 
 
@@ -32,6 +34,9 @@ class Solver:
             halt_on_ampl_error: If True, halts on AMPL errors.
             ipopt_maxiter: Maximum number of iterations for IPOPT.
             ipopt_output_file: File to write IPOPT output to.
+
+        Raises:
+            SolverException if solver status is not "OK"
         """
 
         # Create an IPOPT solver instance
@@ -45,10 +50,7 @@ class Solver:
         # Solve the model
         results = opt.solve(concrete_model, tee=verbose, symbolic_solver_labels=True)
 
-        # Check the solver status
-        self.status = results.solver.status
-        if self.status != SolverStatus.ok:
-            raise Exception(f"Solver did not exit with status OK: {self.status}")
+        self._log_status(results)
 
         return results
 
@@ -60,8 +62,46 @@ class Solver:
             concrete_model: The Pyomo ConcreteModel to solve.
             neos_email: Email address for NEOS server.
             solver_name: Name of the solver to use on NEOS (default is 'ipopt').
+
+        Raises:
+            SolverException if solver status is not "OK"
         """
         os.environ["NEOS_EMAIL"] = neos_email
         solver_manager = SolverManagerFactory("neos")
         results = solver_manager.solve(concrete_model, opt=solver_name)
+
+        self._log_status(results)
+
         return results
+
+    def _log_status(self, results):
+        """
+        Logs the status of the solver results.
+        If status is not OK, raises a SolverException
+        """
+        status = results.solver.status
+
+        logger.info("Status: {}".format(status))
+
+        self.status = status
+
+        if status != SolverStatus.ok:
+            if status == SolverStatus.warning:
+                warning_message = "MIMOSA did not solve succesfully. Status: {}, termination condition: {}".format(
+                    status, results.solver.termination_condition
+                )
+                logger.error(warning_message)
+                raise SolverException(warning_message, utils.MimosaSolverWarning)
+            if status != SolverStatus.warning:
+                raise SolverException(f"Solver did not exit with status OK: {status}")
+
+
+###########################
+##
+## Utils
+##
+###########################
+
+
+class SolverException(Exception):
+    """Raised when Pyomo solver does not exit with status OK"""
